step2_llama_finetune/requirements.txt
# Core ML logic
transformers
peft
trl
datasets
accelerate
bitsandbytes

# Data & Utilities
pandas
tqdm
Pillow
easyocr
bert-score
